adding conditional existence of input into output folder
adding conditional existence of input into output folder

+++ bash_scripts/add_ncbi_annotation.sh
@@ -172,8 +172,19 @@ seconds=$(date "+%Y%m%d_%H%M%S")
 # mkdir add_ncbi_annotation_$sec
 
 cd ${o_DIR}/${output_folder}
-cp ${a_DIR}/${stringtie_input} ${o_DIR}/${output_folder}
 
+if [ -f ${o_DIR}/${output_folder}/$stringtie_input]; then
+    echo "$stringtie_input file found in output directory. Continue."
+    echo ""
+    : 
+else
+    echo "copying $stringtie_input file into the output directory:"
+    cp ${a_DIR}/${stringtie_input} ${o_DIR}/${output_folder}
+    done
+    ###
+fi
+
+# cp ${a_DIR}/${stringtie_input} ${o_DIR}/${output_folder}
 # cd add_ncbi_annotation_$sec
 
 dir1=$(cd -P -- "$(dirname -- "$0")" && pwd -P)
